trading_bot/requirements.txt
python-binance>=1.0.16
python-dotenv>=0.21.0
